Moving some things around, batching seems to work
