+++ index.html
@@ -14,8 +14,8 @@
     <meta name="generator" content="BBEdit 10.5" />
 
     <!-- define our shaders -->
-    <script id="shader-fs" type="x-shader/x-fragment" src="shaders/test_fs.glsl"></script>
-    <script id="shader-vs" type="x-shader/x-vertex" src="shaders/test_vs.glsl"></script>
+    <script id="shader-fs" type="x-shader/x-fragment" src="/shaders/test_fs.glsl"></script>
+    <script id="shader-vs" type="x-shader/x-vertex" src="/shaders/test_vs.glsl"></script>
 
     <!-- main module with runtime -->
     <script type="module" src="js/app.mjs"></script>
